Move min/max to header and rework with BitBlockCounter
